incorporacion de graficos
