+++ Client/src/components/Button/Button.jsx
@@ -1,3 +1,5 @@
+/* eslint-disable no-unused-vars */
+/* eslint-disable react/prop-types */
 import Styles from "./Button.module.css";
 
 const Button = ({ text, onClick, string }) => {
