+++ Client/src/views/Dashboard/Graficos.module.css
@@ -1,0 +1,26 @@
+/* Graficos*/
+.graficoContainer {
+    width: 100%;
+    height: 90%;
+    display: grid;
+    grid-template-columns: 15% 85%;
+    align-items: center;
+    justify-content: center;
+    padding: 1.5rem;
+  }
+  
+  .graficoCharts {
+    display: grid;
+    width: 500px;
+    height: 300px;
+    grid-template-columns: 1fr 1fr;
+    gap: 20px;
+    margin: 5, 30, 20, 5;
+  }
+  
+  @media screen and (max-width: 768px) {
+    .graficoCharts {
+      grid-template-columns: 1fr;
+      margin-top: 30px;
+    }
+  }
